adding estim cylinder using estim_tip_coordinates
the UpdateME.py file now take the voxel coordinates of the estim tip. Brainrender will now add a green sphere at that coordinate, and add a cylinder representing the electrode running from that coordinate to the brain surface. created a vector calculations.py file that will calaculate the different 3D-space measurments of cell distances from the estim_tip_coordinate, and return histograms of those distances, with bin sizes of 100