added new variables in UpdateME.py
can change the radius of the estim shank, and the estim tip. you can specify a gray sphere to load around the estim tip to represent propigration distance in um. cell count is added to the label for each brain region, and (manually added) is placed next to the brain regions you manually added.