Partially working estim to brain region option
there is now a variable called brain_region_estim. if it is set to True than an estim renders in the brain region selected, if False then the render used [x,y,z] coordinates to render the estims location.

issue is right now it renders in the middle of the two hemispheres, not in the center of the left or right side