Create_your_own_manually_curiated_cell_classification
this allows you to manually curriate the cells and non-cells in napari for the cellfinder file (cell_classification.xml). by using the csv file of that, and adding a coulmn titles (Manual), and adding a 2 for cells and a 1 for non-cells. you can then save out that csv file from google sheets, or whereever, and run it through this file. it will create a xml that can load into napari, will new cells and non-cell coordinates. you can then use those two layers in the plugins/Curation tab to create your own training data sets.